+++ uniffi_bindgen/src/bindings/kotlin/gen_kotlin/custom.rs
@@ -3,16 +3,17 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use super::CodeType;
-use crate::ComponentInterface;
+use crate::{interface::DefaultValue, ComponentInterface};
 
 #[derive(Debug)]
 pub struct CustomCodeType {
     name: String,
+    builtin: Box<dyn CodeType>,
 }
 
 impl CustomCodeType {
-    pub fn new(name: String) -> Self {
-        CustomCodeType { name }
+    pub fn new(name: String, builtin: Box<dyn CodeType>) -> Self {
+        CustomCodeType { name, builtin }
     }
 }
 
@@ -23,5 +24,11 @@ impl CodeType for CustomCodeType {
 
     fn canonical_name(&self) -> String {
         format!("Type{}", self.name)
+    }
+
+    fn default(&self, default: &DefaultValue, ci: &ComponentInterface) -> anyhow::Result<String> {
+        self.builtin
+            .default(default, ci)
+            .map_err(|_e| anyhow::anyhow!("Unsupported default value for {}", self.type_label(ci)))
     }
 }