+++ uniffi_bindgen/src/bindings/swift/gen_swift/custom.rs
@@ -7,11 +7,12 @@ use super::CodeType;
 #[derive(Debug)]
 pub struct CustomCodeType {
     name: String,
+    builtin: Box<dyn CodeType>,
 }
 
 impl CustomCodeType {
-    pub fn new(name: String) -> Self {
-        CustomCodeType { name }
+    pub fn new(name: String, builtin: Box<dyn CodeType>) -> Self {
+        Self { name, builtin }
     }
 }
 
@@ -22,5 +23,11 @@ impl CodeType for CustomCodeType {
 
     fn canonical_name(&self) -> String {
         format!("Type{}", self.name)
+    }
+
+    fn default(&self, default: &crate::interface::DefaultValue) -> anyhow::Result<String> {
+        self.builtin
+            .default(default)
+            .map_err(|_e| anyhow::anyhow!("Unsupported default value for {}", self.type_label()))
     }
 }